Crate acceptance_tests_ghidra

This crate contains acceptance tests using Ghidra as a backend for the cwe_checker.

Structs

CweTestCase
A test case containing the necessary information to run an acceptance test.

Constants

ARCHITECTURES
CPU architectures contained in the test samples

COMPILERS
Compilers contained in the test samples

WINDOWS_ARCHITECTURES
CPU architectures for the Windows-based test samples

WINDOWS_COMPILERS
Compilers used for the Windows-based test samples

Functions

all_test_cases
Return a list of all possible test cases (Linux and Windows) for the given CWE.

linux_test_cases
Return a list with all possible Linux test cases for the given CWE.

mark_architecture_skipped
Mark test cases using the given CPU architecture as skipped.

mark_compiler_skipped
Mark test cases using the given compiler as skipped.

mark_skipped
Mark test cases using the given CPU architecture + compiler combination as skipped.

new_test_cases
Generate test cases for all combinations of CPU architecture and compiler given.

print_errors
Print the error messages of failed checks. The error_log tuples are of the form (check_filename, error_message).

windows_test_cases
Return a list with all possible Windows test cases for the given CWE